backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-html==4.1.1
allure-pytest==2.13.2
requests==2.31.0
playwright==1.40.0
locust==2.16.1
jinja2==3.1.2
python-dotenv==1.0.0
locust==2.16.1
playwright==1.40.0
allure-pytest==2.13.2
pydantic==2.5.0